Consolidate EPRV compliance and dtype policy into data_models/config
Keywords, extensions and dtypes had four homes between them: the installed
rvdata, a byte-identical vendored snapshot, the KPF config CSVs, and two
hand-written re-implementations under tests/. Compliance was asserted against
the snapshot rather than the dependency, so it could not catch drift, and 49 of
its 58 EXCEPTIONS entries served a single class that re-checked where each
keyword is written -- enforcement that belongs in the data models.

Manifests now declare an exact BitDepth (rvdata's 8/16/32/64 vocabulary) in
place of the one-sided MinBitDepth floor, and set_data raises on a mismatch
instead of silently upcasting. Renaming the accessor off _get_min_bit_depth is
load-bearing: rvdata's set_data calls that name, so it now reaches RVDataModel's
None-returning base and its upcast goes inert, leaving _assert_bit_depth as the
only enforcement. Empty arrays stay exempt -- every extension is born
np.array([]), which is 64-bit. Declaring a width rather than a dtype lets
np.bool_ satisfy the 8-bit master MASK rows unchanged.

RV BinTable column widths, which an extension manifest cannot express, move to
config/L4-RV-columns.csv; KPF4 no longer imports LEVEL4_RV_TABLE_COLUMNS. The
column order matches rvdata's, so the empty skeleton is unchanged.

config/rvdata-0.4.0/ is deleted along with its packaging glob and pre-commit
exclusions. test_eprv_compliance.py is replaced by a TestEPRVCompliance class in
each test_data_models_l{0,1,2,4}.py, asserting existence and shape against the
installed release, with the cross-cutting checks in test_data_models_base.py.
The surviving divergences are PVN_#, eight unbuilt L2 extensions, three unbuilt
L4, and ANCILLARY_SPECTRUM's HDU type; everything else matches with no
exception. Master manifest-integrity checks move to the master model tests,
which are not subject to EPRV compliance.

The _dtype_policy constants and the _registry oracle now read config rather than
restating it. Fixtures that wrote float64 into 32-bit extensions -- unchecked
under the old floor -- are corrected to float32; two RV assertions that rested
on float64 rounding noise are restated structurally.

Pure refactor: no array, header or RV values change. The one behavior change is
that a dtype violation raises instead of warning.

+++ tests/regression/test_cross_correlation.py
@@ -394,7 +394,7 @@ def _build_cc_kpf2(berv=None, wave_offsets=None, bjd=None):
         ("RED", NORDER_GREEN, NORDER_RED),
     ]:
         wave = np.stack([rows[start + i][0] for i in range(n)]).astype(np.float64)
-        flux = np.stack([rows[start + i][1] for i in range(n)]).astype(np.float64)
+        flux = np.stack([rows[start + i][1] for i in range(n)]).astype(np.float32)
         for fiber in _FIBERS:
             kpf2.set_data(f"{chip}_{fiber}_WAVE", wave.copy())
             kpf2.set_data(f"{chip}_{fiber}_FLUX", flux.copy())
